pyproject.toml
[project]
name = "my-new-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "google-generativeai>=0.8.4",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.7",
    "openai>=1.64.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.42.2",
    "tensorflow>=2.18.0",
]
